+++ frontend/src/components/editor/cell/cell-actions.tsx
@@ -6,12 +6,12 @@ import { useAtomValue } from "jotai";
 import React, {
   Fragment,
   type PropsWithChildren,
-  useEffect,
   useImperativeHandle,
   useMemo,
   useRef,
   useState,
 } from "react";
+import useEvent from "react-use-event-hook";
 import {
   renderMinimalShortcut,
   renderShortcut,
@@ -66,24 +66,35 @@ const CellActionsDropdownInternal = (
   // store the last focused element so we can restore it when the popover closes
   const restoreFocus = useRestoreFocus();
 
-  useEffect(() => {
-    if (open) {
-      // Scroll the button into view when opening the popover
-      // This can be out of view if the popover and a cell's code is
-      // hidden at the same time.
-      raf2(() => {
-        if (buttonRef.current) {
-          buttonRef.current.scrollIntoView({
-            behavior: "auto",
-            block: "nearest",
-          });
-        }
-      });
-    }
-  }, [open]);
+  const handleScrollIntoView = useEvent(() => {
+    raf2(() => {
+      if (buttonRef.current) {
+        buttonRef.current.scrollIntoView({
+          behavior: "auto",
+          block: "nearest",
+        });
+      }
+    });
+  });
+
+  const handleOpenChange: typeof setOpen = useEvent((nextOpenOrCallback) => {
+    // Scroll the button into view when opening the popover
+    // This can be out of view if the popover and a cell's code is
+    // hidden at the same time.
+    setOpen((prevOpen) => {
+      const nextOpen =
+        typeof nextOpenOrCallback === "function"
+          ? nextOpenOrCallback(prevOpen)
+          : nextOpenOrCallback;
+      if (nextOpen) {
+        handleScrollIntoView();
+      }
+      return nextOpen;
+    });
+  });
 
   useImperativeHandle(ref, () => ({
-    toggle: () => setOpen((prev) => !prev),
+    toggle: () => handleOpenChange((prev) => !prev),
   }));
 
   const content = (
@@ -156,7 +167,7 @@ const CellActionsDropdownInternal = (
 
   if (!showTooltip) {
     return (
-      <Popover open={open} onOpenChange={setOpen}>
+      <Popover open={open} onOpenChange={handleOpenChange}>
         <PopoverTrigger asChild={true}>{children}</PopoverTrigger>
         {content}
       </Popover>
@@ -170,7 +181,7 @@ const CellActionsDropdownInternal = (
   );
 
   return (
-    <Popover open={open} onOpenChange={setOpen}>
+    <Popover open={open} onOpenChange={handleOpenChange}>
       <TooltipRoot delayDuration={200} disableHoverableContent={true}>
         {!open && tooltipContent}
         {/* This creates a warning in React due to nested <button> elements.
